# docs/swagger.yaml
definitions:
...
  models.Post:
    properties:
      authorId:
        type: integer
      authorNick:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      likes:
        type: integer
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: RESTful API developed in Golang, intended to serve as the backend for
    a social networking application
  title: SocialMedia-API
paths:
...
  /posts:
    get:
      description: Retrieve all posts from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new post with the data sent in the request body
      parameters:
      - description: Create Post
        example: '{"title": "string", "content": "string"}'
        in: body
        name: post
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Create a new post
      tags:
      - posts
  /posts/{postId}:
    delete:
      description: Delete a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Delete a post
      tags:
      - posts
    get:
      description: Retrieve a post by its ID from the database
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get a post by ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update a post with the data sent in the request body
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: integer
      - description: Post data
        example: '{"title": "string", "content": "string"}'
        in: body
        name: post
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Update a post
      tags:
      - posts
  /posts/{postId}/dislike:
    post:
      description: Decrement the likes count for a post
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Dislike a post
      tags:
      - posts
  /posts/{postId}/like:
    post:
      description: Increment the likes count for a post
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Like a post
      tags:
      - posts
...
  /users/{userId}/posts:
    get:
      description: Retrieve all posts created by a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get all posts by user
      tags:
      - posts
securityDefinitions:
  Bearer:
    description: Provide the JWT token with prefix 'Bearer ' in the text box.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"